+++ src/main.cpp
@@ -167,7 +167,23 @@ class $modify(CCScheduler) {
 		}
 		auto dt2 = dt;
 		if (uwuBot::catgirl->m_state != state::off) {
-			dt2 = 1.f / 240.f;
+			if (Mod::get()->getSettingValue<bool>("lock-delta")) {
+				dt2 = 1.f / 240.f;
+			}
+
+			std::stringstream ss;
+			ss << std::fixed << std::setprecision(2) << static_cast<float>(Mod::get()->getSettingValue<double>("speedhack"));
+			float speed = std::stof(ss.str());
+
+			if (Mod::get()->getSettingValue<bool>("enable-speedhack")) {
+				dt2 *= speed;
+
+				if (Mod::get()->getSettingValue<bool>("speedhack-audio")) {
+					FMOD::ChannelGroup* channel;
+					FMODAudioEngine::sharedEngine()->m_system->getMasterChannelGroup(&channel);
+					channel->setPitch(speed);
+				}
+			}
 		}
 		CCScheduler::update(dt2);
 	}